+++ packages/stream_chat_flutter/lib/src/message_list_view/message_list_view.dart
@@ -944,6 +944,17 @@ class _StreamMessageListViewState extends State<StreamMessageListView> {
 
   Future<void> _markMessagesAsRead() async {
     if (widget.parentMessage case final parent?) {
+      // If we are in a thread, mark the thread as read immediately.
+      await streamChannel?.channel.markThreadRead(parent.id);
+      return;
+    }
+
+    // Otherwise, mark the channel as read immediately.
+    await streamChannel?.channel.markRead();
+  }
+
+  Future<void> _debouncedMarkMessagesAsRead() async {
+    if (widget.parentMessage case final parent?) {
       // If we are in a thread, mark the thread as read.
       debouncedMarkThreadRead.call([parent.id]);
     } else {
@@ -1479,7 +1490,7 @@ class _StreamMessageListViewState extends State<StreamMessageListView> {
     if (!isInThread && !hasUnread) return;
 
     // Mark messages as read if it's allowed.
-    return _markMessagesAsRead();
+    return _debouncedMarkMessagesAsRead();
   }
 
   void _getOnThreadTap() {